refactor: main.js로 이름 변경 및 모듈화
